participants/101/source/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-pptx
aiohttp
pypdf2
python-docx
pillow
pydantic
python-dotenv